Changed to Int

+++ app/Http/Controllers/Auth/RegisterController.php
@@ -197,13 +197,13 @@ class RegisterController extends Controller
     private function calculateAgeFromId(string $idNumber): int
     {
         // Extract the first two digits for the year of birth (YY)
-        $year = substr($idNumber, 0, 2);
+        $year = (int) substr($idNumber, 0, 2);
 
         // Extract the month of birth (MM)
-        $month = substr($idNumber, 2, 2);
+        $month = (int) substr($idNumber, 2, 2);
 
         // Extract the day of birth (DD)
-        $day = substr($idNumber, 4, 2);
+        $day = (int) substr($idNumber, 4, 2);
 
         // Ensure valid day and month values
         if (!checkdate($month, $day, $year)) {
@@ -211,10 +211,10 @@ class RegisterController extends Controller
         }
 
         // Get the current year in full (YYYY format)
-        $currentYear = date('Y');
+        $currentYear = (int) date('Y');
 
         // Determine if the century is 19xx or 20xx based on the extracted year
-        $year = ($year > date('y')) ? '19' . $year : '20' . $year;
+        $year = ($year > (int) date('y')) ? (int) ('19' . $year) : (int) ('20' . $year);
 
         // Create a DateTime object from the extracted birth date (YYYY-MM-DD format)
         $birthDate = \DateTime::createFromFormat('Y-m-d', "$year-$month-$day");
